debugged of linking functionality

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, ForeignKey, DateTime, Boolean
+from sqlalchemy import BigInteger, Column, Integer, String, Text, ForeignKey, DateTime, Boolean
 from sqlalchemy.orm import relationship, declarative_base
 from datetime import datetime
 
@@ -7,8 +7,8 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = 'users'
     
-    id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True, nullable=False)
+    id = Column(BigInteger, primary_key=True)
+    telegram_id = Column(BigInteger, unique=True, nullable=False)
     name = Column(String, nullable=False)
     summary = Column(Text, nullable=True)
 
@@ -17,9 +17,9 @@ class User(Base):
 class Couple(Base):
     __tablename__ = 'couples'
     
-    id = Column(Integer, primary_key=True)
-    user1_id = Column(Integer, ForeignKey('users.id'))
-    user2_id = Column(Integer, ForeignKey('users.id'))
+    id = Column(BigInteger, primary_key=True)
+    user1_id = Column(BigInteger, ForeignKey('users.id'))
+    user2_id = Column(BigInteger, ForeignKey('users.id'))
     
     user1 = relationship('User', foreign_keys=[user1_id])
     user2 = relationship('User', foreign_keys=[user2_id])
@@ -28,9 +28,9 @@ class Couple(Base):
 class Conversation(Base):
     __tablename__ = 'conversations'
     
-    id = Column(Integer, primary_key=True)
-    couple_id = Column(Integer, ForeignKey('couples.id'))
-    user_id = Column(Integer, ForeignKey('users.id'))
+    id = Column(BigInteger, primary_key=True)
+    couple_id = Column(BigInteger, ForeignKey('couples.id'))
+    user_id = Column(BigInteger, ForeignKey('users.id'))
     message = Column(Text, nullable=False)
     timestamp = Column(DateTime, default=datetime.utcnow)
 
@@ -40,9 +40,9 @@ class Conversation(Base):
 class ScheduledAction(Base):
     __tablename__ = 'scheduled_actions'
     
-    id = Column(Integer, primary_key=True)
-    couple_id = Column(Integer, ForeignKey('couples.id'))
-    user_id = Column(Integer, ForeignKey('users.id'))
+    id = Column(BigInteger, primary_key=True)
+    couple_id = Column(BigInteger, ForeignKey('couples.id'))
+    user_id = Column(BigInteger, ForeignKey('users.id'))
     action = Column(String, nullable=False)
     trigger_time = Column(DateTime, nullable=False)
     is_active = Column(Boolean, default=True)
@@ -53,17 +53,17 @@ class ScheduledAction(Base):
 class UserActionLog(Base):
     __tablename__ = 'user_action_logs'
     
-    id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, ForeignKey('users.id'))
+    id = Column(BigInteger, primary_key=True)
+    user_id = Column(BigInteger, ForeignKey('users.id'))
     action = Column(String, nullable=False)
     timestamp = Column(DateTime, default=datetime.utcnow)
 
 class PendingCouple(Base):
     __tablename__ = 'pending_couples'
 
-    id = Column(Integer, primary_key=True)
-    requester_id = Column(Integer, ForeignKey('users.id'), nullable=False)
-    requested_id = Column(Integer, ForeignKey('users.id'), nullable=True)  # Make nullable
+    id = Column(BigInteger, primary_key=True)
+    requester_id = Column(BigInteger, ForeignKey('users.id'), nullable=False)
+    requested_id = Column(BigInteger, ForeignKey('users.id'), nullable=True)  # Make nullable
     token = Column(String, unique=True, nullable=False)
 
     requester = relationship('User', foreign_keys=[requester_id])
@@ -72,7 +72,7 @@ class PendingCouple(Base):
 class Translation(Base):
     __tablename__ = 'translations'
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     original_text = Column(Text, nullable=False)
     target_language = Column(String, nullable=False)
     translated_text = Column(Text, nullable=False)
